requirements.txt
Flask==2.3.2
python-pptx==0.6.21